+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+1.0.2a
+- Fixed vanilla cocoa beans not growing correctly due to a mixin
+---
 1.0.2
 - Added Spanish translations
 - Added obese cocoa
@@ -6,7 +9,6 @@
 - Chickens can now eat apple seeds
 - Apples can now grow into obese apples
 - Fixed foliage not breaking when the block has been chopped
-
 ---
 1.0.1
 - Added flowering oak trees